chore(pkg): relocate http pool initialisation from utils to oktaApiAuth
